Revise the `CrapsGame` stub to add methods for validating bets in different game states. Add a `Table` subclass: `CrapsTable, which uses methods from `CrapsGame` to validate bets and the validity of all the table's bets. Add unit test for `CrapsTable`.
Change to the `Wheel` class to stop it from building bins with `BinBuilder` on init. This is to prevent subclasses of `Table` from unnecessarily building bins (such as in `CrapsTable`). This required a small change to test_roulette_game.py and test_simulator.py to call `BinBuilder.build_bins` after initiating a `Table` instance.